CCDUIT/requirements.txt
fastapi==0.115.6
flask-cors==4.0.1
openpyxl==3.1.5
paho-mqtt==2.1.0
psutil==6.1.0
pydantic==2.10.3
rdflib==7.1.1
requests==2.32.3
uvicorn==0.32.1
